Move palladium config code out into its own module
